Add @controllers instance variable to Spine.Controller
This change introduces the @controllers instance variable in
Spine.Controller. This keeps track of any controllers that have
been added to an existing controller during a call to append,
prepend, or appendTo. This also resets the controllers variable when
calling html.

The purpose of this is to remove the repetitive pattern of keeping track
of controllers added yourself, for example in the element pattern.